+++ apps/src/openni_feature_persistence.cpp
@@ -82,7 +82,7 @@ class OpenNIFeaturePersistence
     typedef typename Cloud::Ptr CloudPtr;
     typedef typename Cloud::ConstPtr CloudConstPtr;
 
-    OpenNIFeaturePersistence (double &subsampling_leaf_size,
+    OpenNIFeaturePersistence (float &subsampling_leaf_size,
                               double &normal_search_radius,
                               std::vector<float> &scales_vector,
                               double &alpha,
@@ -242,7 +242,7 @@ usage (char ** argv)
     for (unsigned deviceIdx = 0; deviceIdx < driver.getNumberDevices (); ++deviceIdx)
     {
       cout << "Device: " << deviceIdx + 1 << ", vendor: " << driver.getVendorName (deviceIdx) << ", product: " << driver.getProductName (deviceIdx)
-              << ", connected: " << (int)driver.getBus (deviceIdx) << " @ " << (int)driver.getAddress (deviceIdx) << ", serial number: \'" << driver.getSerialNumber (deviceIdx) << "\'" << endl;
+              << ", connected: " << driver.getBus (deviceIdx) << " @ " << driver.getAddress (deviceIdx) << ", serial number: \'" << driver.getSerialNumber (deviceIdx) << "\'" << endl;
       cout << "device_id may be #1, #2, ... for the first second etc device in the list or" << endl
            << "                 bus@address for the device connected to a specific usb-bus / address combination (works only in Linux) or" << endl
            << "                 <serial-number> (only in Linux and for devices which provide serial numbers)"  << endl;
@@ -265,7 +265,7 @@ main (int argc, char **argv)
   }
 
   // Parse arguments
-  double subsampling_leaf_size = default_subsampling_leaf_size;
+  float subsampling_leaf_size = default_subsampling_leaf_size;
   pcl::console::parse_argument (argc, argv, "-octree_leaf_size", subsampling_leaf_size);
   double normal_search_radius = default_normal_search_radius;
   pcl::console::parse_argument (argc, argv, "-normal_search_radius", normal_search_radius);
